Update ember-table to be compatible with EmberJS rc3
In emberjs/ember.js@9014b5cebab5037bef141ab9599a7fc7bab62178 there was an update to remove the defaultContainer global instantiation.

This broke ember-table because the LazyContainerView is created using the .create() method instead of view.createChildView().  The createChildView method will instantiate all of the container objects (and other childView *stuff*), whereas create() does not.

After this change, ember-table will be rc3 compatible.

Thanks to @stefanpenner for assisting in diagnosis.