Added minimum jQuery UI build info to site
The entire jQuery UI package weighs in at 228k after minifcation. The modules that are required for ember-table are only 55k after min. After digging through the code, it appears that only the following modules are required:

* core
* widgets
* mouse
* resizable
* sortable

So it would probably be helpful to include this info in the docs to avoid users downloading the entire package when it isn't necessary.