Simpler column resizing when forceFill is on.
Resizing a column in forceFill mode will now set the column's
canAutoResize property to true. Force fill will then fill remaining
space by adjusting any columns which canAutoResize.

If you try to resize the last (or only) column for which canAutoResize
is true, it will snap back to its original size. This is reasonable
behavior because you can still adjust it by resizing the other columns,
but the forceFill behavior is preserved.