Fix lazy container bug
When changing content of ember-table I end up with an error stating
that I can't call get on undefined. This stems from a problem with
lazy container where the container views length is longer than the
actual number of views. I'm unsure if this is an ember problem or if
its something to do with the implementation of lazy container itself.

It might be that you can't rely on ember having created all the views.
Checking to see if the view exists before settings its content fixes
this and I have had no noticeable problems with this extra check.