Fix undefined table component
There is a race condition where if a header cell is removed from the dom during
resizing, tableComponent becomes undefined in elementSizeDidChange because the
associated view has already been destroyed by the time the afterRender portion
of the run loop happens. This throws an exception. This commit properly cancels
the scheduled callback when the view is destroyed, preventing the exception.