Add virtual root (#10)
In current behavior, we have root of linked list tree as an element to display in the table. This assumes that the table has one row that is the ancestor of every other row and this is how the demo app mocks its data.

However, in many table, we do not have one row at the top level but multiple rows at top level. To handle this case, we should have a root node that's the parent of all top level row but does not display in the table. This PR adds a virtual root to the linked list tree to support that.