[refactor] Use position: sticky; for fixed headers/footers/columns
This PR switches us to a rendering strategy using position: sticky;
in all modern browsers. This is much more performant and simpler in the
code, and maintains semantic table structure - everything we've been
after!

For IE11, the only browser we support which does not support sticky, we
use a rudimentary polyfill. The polyfill is not comprehensive, it only
solves our exact use case by using `translateX/Y` on scroll.