[bugfix] Fix fixed column behavior
Even with the changes to decorators upstream there, we still can't
decorate a property without modifying the value on the prototype. In
this case it ends up manually defining an undefined value, which breaks
the prototype chain. We'll have to wait for decorators v2 transform to
see if things get fixed.