[refactor] Combine ember-table and tree-table
Based on the recent work on CollapseTree we now have a cheap data
structure which can easily wrap a top level array. This means we can
combine {{tree-table}} with {{ember-table}}, with the former "plain
array" functionality of {{ember-table}} handled as the case where we
have a tree with no children on any nodes.

There are more optimizations that'll go into this work - in particular
I think we can clean up the selection flows a bit for tree tables.
For now, this is consolidation pre-major refactor.