[polyfill] Polyfill the behavior of position: sticky; in thead/tfoot (#507)
The previous implementation of fixed headers/footers wouldn't actually
work for headers and footers that have more than 1 row because it fixed
td/th elements directly due to a bug in the Edge/Chrome versions of
sticky. This PR adds a polyfill which does several things:

* It positions the headers/footers correctly individually based
on a measurement of the row height
* It repositions whenever the row height changes by using a ResizeSensor
* It repositions whenever cells or rows are added by using
MutationObservers

This is definitely suboptimal, but it's much better than our previous
setup and should be temporary.

The polyfills have been moved into private, and the legacy polyfill has
been refactored to match the style of the new one. They're both class
based now and use a WeakMap to tie themselves directly to an element
rather than the component class itself. This means we don't dirty the
component (or modify its hidden class) and we and we keep all of our
implementation details completely private.

One final change was to add an explicit width on the table itself. The
previous hack of setting `width: 0;` doesn't actually work in IE11, it
collapses the entire table, so we need to actually set it to the real
width (as defined by the columns).