[feat] ColumnTree
This PR adds adds a ColumnTree data structure which manages the
hierarchy of columns and subcolumns. This allows us to have a nested
tree of columns with any configuration. All levels of columns can be
reordered and resized, including fixed columns on the left and the
right hand side.

Documentation has been added to the docs site for the new features, and
they are fully tested.