[bugfix] Avoid clickable elements when toggling row selection (#538)
Currently, users will have to manually handle events in their usage of
Ember Table when clicking a button or other interactable element to
prevent row selection from toggling. This PR causes the table to ignore
any clicks that originate in a clickable element (`button`, `label`,
`input`, `a`, or `select` for nows).