[refactor] Refactors Sorting Markup (#540)
This PR refactors the Sorting markup to not use a button and instead rely on
the table headers directly. This is easier for us to style, and more flexible
with yielding. It also removes the dropdownActions API and the dropdown
associated with it. Dropdown actions are very unique to the application, so
it makes more sense for users to implement themselves with yielding.