[feat] Separate row/checkbox selection, fix reordering/resizing (#544)
This PR does several things:

1. Separates row and checkbox selection settings. Users can now
independently choose the how selection works with checkboxes or by
clicking on rows. This creates a matrix of options that allows a number
of different UX patterns with tables.

2. Fixes the behavior of reordering. Before, the final drop position was
determined by the position of the middle of the resize indicator. This
was counterintuitive since the reorder indicator could be in a
completely different place than the cursor, and it actually did not work
at all in cases where the header was larger than the column that was
attempting to be reordered because headers would not travel beyond the
confines of the table. Now, the final drop position is determined by the
cursor, and the reorder indicator is always centered on the cursor.

3. Fixes the resizing. Before, users could begin a resize and end it
without removing the `is-resizing` class, locking the table into having
the resizing cursor forever.