[bugfix] Fix Column Sorting (#554)
It was possible for columns to get into a bad state based on the
ordering of Hammer event handlers and the click event handler. This PR
fixes it temporarily, but I think this needs further refactoring.

Testing this currently would be very difficult (it would require a lot
of intricate ordering of events) so this PR does not add tests. Locking
us into a series of events that will likely be refactored seems
ill-advised.

Fixes #553