use window scroll position for ember-router-scroll
This configures the ember-router-scroll addon to properly remember
scroll position between routes. This is installed as part of
ember-cli-addon-docs, but its config in the dummy app here pointed to an
element that doesn't exist. Instead I just removed the config and it now
uses the `window` scroll position instead, which works well.