Row meta computes index from _lastKnownIndex and _prevSiblingMeta.index
When computing a row's meta index, the last known index (i.e., from last accessed) is reconciled with its previous sibling's index. For example, if its previous sibling's index is the same as its own, this assumes a shift is needed as a result of a lower index insertion. This then cascades forward.