[bugfix] Only clean caches on table destroy (#611)
Currently, whenever rows change we attempt to erase the row meta values
for rows which are no longer part of the table. This is an O(n)
operation which is fairly expensive, and we were doing it wrong as we
were not traversing children rows at all.

This PR moves the cache cleaning to only occur when the table is fully
destroyed. This means that we may hold onto meta objects longer than is
ideal, but we also won't incur an expensive operation every time the
table is updated. The alternative would be to fix the cache cleaning so
that we traverse child rows as well.