"Disable collapse" support at a row level (#629)
In this commit I make it possible to disable collapsing at a row level rather than just at the total tree level. I need this functionality to allow my tables to "hide" the children below a certain level but I thought this might be a better approach than a maxDepth because this is more flexible and allows more user choice about what they want to build.