Replace `constructor` with `init` (#633)
* Replace `constructor` with `init`

As per [RFC 337](https://github.com/emberjs/rfcs/pull/337), the way
classes are initialized has changed. Namely, arguments are no longer
available in `constructor` but they are still available in `init`.

Because we used `constructor` to set properties from passed in arguments
before, we started seeing error on Travis:

```
Error: Property set failed: object in path "unwrappedApi" could not be found.
```

This change switches to using `init` where appropriate and is a
backwards compatible change.

* Move class properties initialisation into `init`

An unfortunate side-effect of replacing `constructor` with `init` is
that class properties are no longer assigned properly. They used to be
assigned after `init`. So all class properties initializers that are not decorated
with `@argument` need to be moved in `init`.