[bugfix] Recompute row meta index when previous prepend causes shift (#623)
* Failing test for stale rowMeta index when prepending row

* Row meta computes index from _lastKnownIndex and _prevSiblingMeta.index

When computing a row's meta index, the last known index (i.e., from last accessed) is reconciled with its previous sibling's index. For example, if its previous sibling's index is the same as its own, this assumes a shift is needed as a result of a lower index insertion. This then cascades forward.

* Update prev and next to use _lastKnownIndex

* Initialize _lastKnownIndex with null value

* Clarify "rowMeta index works" test name