[PERF] Set default bufferSize to 1
By default, ember-table allocates a buffer of 20 rows
before and after the visible ones.

This default value forces a lot of computation in the case of tables with many columns and complex cells,
from syncinc more data (from rows to cells for example) and triggering more observers
(based on the number of computed properties defines per cell for example).
VC also does not allocate the part of the buffer that should be above the visible first row,
as there is none when loading the table. This means that the initial scroll will also
lead to the creation of 20 rows in the DOM (and the underlying computation to render the content).

From a pure UI/UX point of view, we don't need more than 1 row on each side of the table,
and it's already the default for `vertical-collection`.