[TEST] Add rowCount to page object for thead, tbody and tfoot
In some tests, it is interesting to see if a row is properly inserted or removed.
Because `vertical-collection` tries to not render all the rows, we can't simply
count the number of `tr`. Instead we use the `data-test-row-count` attribute
to keep track of the number of available rows in a table.