DOCS: Test that subcolumns docs render cells
Fix rendering of generated dummy rows. Cells with valuePaths longer than
a single character were being skipped over in the dummy row's `unknownProperty`
hook.
Fix by removing the length check in `unknownProperty` and explicitly adding
the needed properties to the class definition so that the only accesses
that hit `unknownProperty` should be `valuePath`s for rendering.