Use objectAt to fetch rowValues (cause meta alloc)
Although the `selection` set contains raw values only, some other code
(like the summing of selected counts for group selection state in
`collapse-tree.js`) expect that all items in the `selection` set have a
meta allocated.

The non-meta-allocated rows were added where the bare `children`
property of a rowValue was being iterated. Instead with this change the
`tree` is referenced for pulling out the children, meaning the meta
cache system is exercised and metas allocated for any selected row.