[DOCS] Miscellaneous small fixes
Remove undefined `py-2` and `pr-4` CSS classes (vestiges from when tailwind
was included in the build from addon-docs).
Consolidate example options to use `demo-options` class wherever possible.
Remove unneeded `demo-option` class.

Fix the 'color' property on the columns in the custom-header example so that
the correct bg-color CSS classes are applied.

Use a separate selection property `demoSelection` for the last demo on the
'row selection' page. When the `selection` property was shared amongst all
the demos on the page, it was possible to select a row in one of the demos
that didn't exist in the later ones, and would cause an error.

Fix typos.