[TEST] Use non closure actions in 1.12 and closure
When tests are run with the 1.12 try scenario it will use non
closure actions otherwise it will use closure actions.

I had to compile the tempalte string at test run time and not at
build time because i wanted to keep the same string instead of
duplicating the string.

As long as 1.12 is in the try this will test both cases, but if 1.12
is removed then non closure actions wont be tested.