Add "slack" width constraint modes (#865)
Adds two width constraint modes: `eq-container-slack` and `gte-container-slack`. These build on their non-"slack" counterparts by allocating excess container whitespace into an additional column that vanishes when it is no longer needed. This resembles behavior in applications like Excel and Google sheets.

Also adds `initialFillMode` parameter that is used exclusively in slack modes to perform the initial column layout. Values and default are the same as `fillMode`. In `eq-container-slack mode`, `fillMode` and `initialFillMode` can be set to different values. The former is used for enforcing the width constraint, while the latter is used only for initial sizing. Examples in docs.