Enforce min and max width constraints before initial fill (#881)
Ensures that min and max column widths are enforced on initial render.

Before, min and max column widths were only enforced in two cases:
* when the widthConstraint mode was violated
* when the user manually resized column

In particular, if `widthConstraint` was `none`, or the chosen `widthConstraint` was not initially violated (e.g. `lte-container` with few columns), min and max column widths would be ignored on initial render.