Detect row count test selector to enable observers (#893)
These observers were installed based on `runInDebug`. As
ember-test-selectors uses its own configuable logic for when to strip
test selectors, the two systems of "debug" mode can get out of sync.

For example, if you build Ember in a "test" environment but with a
production Ember.js asset, then `runInDebug` is a no-op despite the test
selectors being in place. You may configure Ember CLI this way if you
want to ensure your application tests run in a mode similar to
production.

Swap them to detect the attribute binding presence, which should not be
there when ember-test-selectors is stripping selectors/bindings.