Refactor scale via offsetHeight use precise API
offsetHeight returns an integer, making it a poor API to base the
determination of element scale on. Instead, use an alternative and
precise API. Only allow a reasonable amount of precision by decimal
place, and for cases where the scale number is reasonably rounded do so.

The offsetHeight method includes padding in returned sizes, where
getComputedStyle does not. Check them against each other, and if they
diverge signifcantly (>1) throw an exception.

To make ember-table itself compatible with this appraoch, pass the
`scale` value around the system a bit more instead of trying to derive
it in many places.